temp menu, level loader
Added a temp menu to create buttons for different scenes we want to try. Also a level loader with a loading screen. The temp menu is loaded when pressing Play on the main menu. Temp art.
