+++ Assets/Scenes/Menus/TempStartMenu.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a88d23d8ffe37584497d6c65049a0d80
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
