+++ Assets/Scripts/MenuScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e8da3fc5113d18f4d85162537e27a953
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
